CIFAR10-object-recognition/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.1.19=hecd8cb5_1
certifi=2020.12.5=py38hecd8cb5_0
cycler=0.10.0=py38_0
ffmpeg=4.3=h0a44026_0
freetype=2.10.4=ha233b18_0
gettext=0.20.2=h7535e17_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.5=h91ad68e_1002
icu=58.2=h0a44026_3
intel-openmp=2019.4=233
jpeg=9b=he5867d9_2
kiwisolver=1.3.1=py38h23ab428_0
lame=3.100=h1de35cc_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libiconv=1.16=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
libxml2=2.9.10=h7cdb67c_3
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.3=h23ab428_0
matplotlib=3.3.4=py38hecd8cb5_0
matplotlib-base=3.3.4=py38h8b3ea08_0
mkl=2019.4=233
mkl-service=2.3.0=py38h9ed2024_0
mkl_fft=1.3.0=py38ha059aab_0
mkl_random=1.1.1=py38h959d312_0
ncurses=6.2=h0a44026_1
nettle=3.4.1=h3018a27_0
ninja=1.10.2=py38hf7b0b51_0
numpy=1.19.2=py38h456fd55_0
numpy-base=1.19.2=py38hcfb5961_0
olefile=0.46=py_0
openh264=2.1.0=hd9629dc_0
openssl=1.1.1k=h9ed2024_0
pillow=8.2.0=py38h5270095_0
pip=21.0.1=py38hecd8cb5_0
pyparsing=2.4.7=pyhd3eb1b0_0
python=3.8.8=h88f2d9e_4
python-dateutil=2.8.1=pyhd3eb1b0_0
pytorch=1.8.1=py3.8_0
readline=8.1=h9ed2024_0
setuptools=52.0.0=py38hecd8cb5_0
six=1.15.0=py38hecd8cb5_0
sqlite=3.35.4=hce871da_0
tk=8.6.10=hb0a8c7a_0
torchaudio=0.8.1=py38
torchvision=0.9.1=py38_cpu
tornado=6.1=py38h9ed2024_0
typing_extensions=3.7.4.3=pyha847dfd_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
